+++ src/components/Node/Node.module.scss
@@ -38,7 +38,7 @@
   color: $black;
 }
 
-.ExpandedNode {
+.expandedNode {
   position: absolute;
   display: block;
   top: $collapsedNodeHeight;
