+++ src/styles/_nodes.scss
@@ -4,7 +4,7 @@
   min-width: 400px;
   z-index: 1;
 
-  .ExpandedNode {
+  .expandedNode {
     color: $black;
   }
